+++ src/main/java/com/tiviacz/travelersbackpack/network/ServerboundSlotPacket.java
@@ -18,18 +18,18 @@ import net.minecraft.world.item.ItemStack;
 import java.util.ArrayList;
 import java.util.List;
 
-public record ServerboundSlotPacket(byte selectType, List<Integer> unsortables,
+public record ServerboundSlotPacket(int selectType, List<Integer> unsortables,
                                     List<Pair<Integer, Boolean>> memorizedSlots) implements CustomPacketPayload {
     public static final Type<ServerboundSlotPacket> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(TravelersBackpack.MODID, "slots"));
     public static final StreamCodec<FriendlyByteBuf, ServerboundSlotPacket> STREAM_CODEC = StreamCodec.composite(
-            ByteBufCodecs.BYTE, ServerboundSlotPacket::selectType,
+            ByteBufCodecs.INT, ServerboundSlotPacket::selectType,
             ByteBufCodecs.INT.apply(ByteBufCodecs.list()), ServerboundSlotPacket::unsortables,
             ByteBufCodecs.fromCodec(Codec.mapPair(Codec.INT.fieldOf("index"), Codec.BOOL.fieldOf("matchComponents")).codec()).apply(ByteBufCodecs.list()), ServerboundSlotPacket::memorizedSlots,
             ServerboundSlotPacket::new
     );
 
-    public static final byte UNSORTABLES = (byte)0;
-    public static final byte MEMORY = (byte)1;
+    public static final int UNSORTABLES = 0;
+    public static final int MEMORY = 1;
 
     public static void handle(final ServerboundSlotPacket message, ServerPlayNetworking.Context ctx) {
         ctx.player().getServer().execute(() -> {
